fix(tja2osu): fix wrong note timing when `#MEASURE` is applied on re-measured measure with non-1.0 `#SCROLL`
This is due to `#MEASURE` only updates the last red timing point,
but the last timing point is green for re-measured measure with non-1.0 `#SCROLL`
(2 timing points (a red and then a green) are added for each such measure),
so the note timing calculated from the green timing point will be wrong.

To solve this issue, the note timing is now calculated
always from the red timing point instead.